blockify-distillation-service/requirements.txt
# Blockify Distillation Service
# Production dependencies

# Web framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# HTTP client
requests==2.31.0
httpx==0.26.0

# ML/Embeddings
numpy==1.26.3
scikit-learn==1.4.0
faiss-cpu==1.7.4

# Graph algorithms
networkx>=3.2

# Database backends
sqlalchemy==2.0.25
aiosqlite==0.19.0  # SQLite async support
# Optional: uncomment for additional backends
# asyncpg==0.29.0  # PostgreSQL
# redis==5.0.1     # Redis

# Observability
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-exporter-otlp==1.22.0

# Logging
structlog==24.1.0

# Configuration
python-dotenv==1.0.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development
black==24.1.0
ruff==0.1.14
mypy==1.8.0
